CategoryCell - свойства класса сделаны приватными, для работы со свойствами сделаны интернал методы
