+++ Tracker/ViewControllers/CategorySelect/CategoryCell.swift
@@ -3,9 +3,11 @@ import SnapKit
 
 class CategoryCell: UITableViewCell {
     
+    //MARK: - Public Properties
+    
     static let reuseIdentifier = "CategoryCell"
     
-    lazy var categoryLabel: UILabel = {
+    private lazy var categoryLabel: UILabel = {
         let label = UILabel()
         label.font = .systemFont(ofSize: 17, weight: .regular)
         label.textColor = .trackerBlack
@@ -13,13 +15,15 @@ class CategoryCell: UITableViewCell {
         return label
     }()
     
-    lazy var checkmarkImageView: UIImageView = {
+    private  lazy var checkmarkImageView: UIImageView = {
         let imageView = UIImageView()
         imageView.image = UIImage(systemName: "checkmark")
         imageView.contentMode = .center
         imageView.isHidden = true
         return imageView
     }()
+    
+    //MARK: - Public Methods
     
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
@@ -34,6 +38,12 @@ class CategoryCell: UITableViewCell {
         categoryLabel.text = categoryName
         checkmarkImageView.isHidden = !isSelected
     }
+    
+    func isSelected(selected: Bool) {
+        checkmarkImageView.isHidden = !selected
+    }
+    
+    //MARK: - Private Methods
     
     private func setUI() {
         backgroundColor  = .trackerBackground
