+++ Tracker/ViewControllers/FiltersViewController/FiltersCell.swift
@@ -72,5 +72,4 @@ final class FiltersCell: UITableViewCell {
         }
     }
     
-    
 }
