+++ Tracker/ViewControllers/StatisticsViewController/StatisticsCounterView.swift
@@ -1,7 +1,9 @@
 import UIKit
 import SnapKit
 
-class StatisticsCounter: UIView {
+final class StatisticsCounter: UIView {
+    
+    //MARK: - Public Properties
     
     var counter: Int {
         didSet {
@@ -14,6 +16,8 @@ class StatisticsCounter: UIView {
             counterLabel.text = counterName
         }
     }
+    
+    //MARK: - Private Properties
     
     private lazy var counterLabel: UILabel = {
         let label = UILabel()
@@ -43,6 +47,8 @@ class StatisticsCounter: UIView {
         return stack
     }()
     
+    //MARK: - Public Methods
+    
     init(counter: Int, counterName: String) {
         self.counter = counter
         self.counterName = counterName
@@ -53,6 +59,13 @@ class StatisticsCounter: UIView {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
+    
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        addGradientBorder(width: 1)
+    }
+    
+    //MARK: - Private Methods
     
     private func setupView() {
         layer.cornerRadius = 16
@@ -73,36 +86,7 @@ class StatisticsCounter: UIView {
         }
     }
     
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        addGradientBorder(width: 1)
-    }
 }
 
-extension UIView {
-    func addGradientBorder(width: CGFloat = 1) {
-        let gradientLayer = CAGradientLayer()
-        //по rgb из фигмы
-        gradientLayer.colors = [
-            UIColor(red: 0.0, green: 0.482, blue: 0.98, alpha: 1.0).cgColor,
-            UIColor(red: 0.275, green: 0.902, blue: 0.616, alpha: 1.0).cgColor,
-            UIColor(red: 0.992, green: 0.298, blue: 0.286, alpha: 1.0).cgColor
-        ]
-        
-        gradientLayer.startPoint = CGPoint(x: 1, y: 0.5)
-        gradientLayer.endPoint = CGPoint(x: 0, y: 0.5)
-        gradientLayer.frame = bounds
-        
-        let borderLayer = CAShapeLayer()
-        let insetBounds = bounds.insetBy(dx: width, dy: width)
-        borderLayer.path = UIBezierPath(roundedRect: insetBounds, cornerRadius: layer.cornerRadius).cgPath
-        borderLayer.lineWidth = width
-        borderLayer.fillColor = UIColor.clear.cgColor
-        borderLayer.strokeColor = UIColor.black.cgColor
-        gradientLayer.mask = borderLayer
-        
-        layer.addSublayer(gradientLayer)
-    }
-}
